refactor(ng-dev/pr): move pr configuration out of the merge directory
As part of #203, refactoring the pr directory into more reusable pieces.
