+++ ng-dev/pr/config.ts
@@ -6,15 +6,21 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ConfigValidationError, GithubConfig} from '../../utils/config';
-
-import {GithubApiMergeStrategyConfig} from './strategies/api-merge';
+import {ConfigValidationError, GithubConfig} from '../utils/config';
 
 /**
  * Possible merge methods supported by the Github API.
  * https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button.
  */
 export type GithubApiMergeMethod = 'merge' | 'squash' | 'rebase';
+
+/** Configuration for the Github API merge strategy. */
+export interface GithubApiMergeStrategyConfig {
+  /** Default method used for merging pull requests */
+  default: GithubApiMergeMethod;
+  /** Labels which specify a different merge method than the default. */
+  labels?: {pattern: string; method: GithubApiMergeMethod}[];
+}
 
 /**
  * Configuration for the merge script with all remote options specified. The
