+++ ng-dev/pr/merge/BUILD.bazel
@@ -16,6 +16,7 @@ ts_library(
     ],
     deps = [
         "//ng-dev/commit-message",
+        "//ng-dev/pr:config",
         "//ng-dev/pr/common",
         "//ng-dev/release/config",
         "//ng-dev/release/versioning",
